docs/coverage/node/html/test/wpt/xhr/historical.js.gcov.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">

<html lang="en">

<head>
  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
  <title>LCOV - lcov.info - test/wpt/xhr/historical.js</title>
  <link rel="stylesheet" type="text/css" href="../../../gcov.css">
</head>

<body>

  <table width="100%" border=0 cellspacing=0 cellpadding=0>
    <tr><td class="title">LCOV - code coverage report</td></tr>
    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>

    <tr>
      <td width="100%">
        <table cellpadding=1 border=0 width="100%">
          <tr>
            <td width="10%" class="headerItem">Current view:</td>
            <td width="35%" class="headerValue"><a href="../../../index.html">top level</a> - <a href="index.html">test/wpt/xhr</a> - historical.js<span style="font-size: 80%;"> (source / <a href="historical.js.func-sort-c.html">functions</a>)</span></td>
            <td width="5%"></td>
            <td width="15%"></td>
            <td width="10%" class="headerCovTableHead">Hit</td>
            <td width="10%" class="headerCovTableHead">Total</td>
            <td width="15%" class="headerCovTableHead">Coverage</td>
          </tr>
          <tr>
            <td class="headerItem">Test:</td>
            <td class="headerValue">lcov.info</td>
            <td></td>
            <td class="headerItem">Lines:</td>
            <td class="headerCovTableEntry">41</td>
            <td class="headerCovTableEntry">41</td>
            <td class="headerCovTableEntryHi">100.0 %</td>
          </tr>
          <tr>
            <td class="headerItem">Date:</td>
            <td class="headerValue">2025-01-23 03:02:29</td>
            <td></td>
            <td class="headerItem">Functions:</td>
            <td class="headerCovTableEntry">2</td>
            <td class="headerCovTableEntry">2</td>
            <td class="headerCovTableEntryHi">100.0 %</td>
          </tr>
          <tr>
            <td></td>
            <td></td>
            <td></td>
            <td class="headerItem">Branches:</td>
            <td class="headerCovTableEntry">3</td>
            <td class="headerCovTableEntry">3</td>
            <td class="headerCovTableEntryHi">100.0 %</td>
          </tr>
          <tr><td><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
        </table>
      </td>
    </tr>

    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
  </table>

  <table cellpadding=0 cellspacing=0 border=0>
    <tr>
      <td><br></td>
    </tr>
    <tr>
      <td>
<pre class="sourceHeading">           Branch data     Line data    Source code</pre>
<pre class="source">
<a name="1"><span class="lineNum">       1 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          1 : import assert from 'node:assert/strict';</span></a>
<a name="2"><span class="lineNum">       2 </span>                :<span class="lineCov">          1 : import { XMLHttpRequest } from '../../../lib/whatwg-xhr.js';</span></a>
<a name="3"><span class="lineNum">       3 </span>                :<span class="lineCov">          1 : </span></a>
<a name="4"><span class="lineNum">       4 </span>           [<span class="branchCov" title="Branch 0 was taken 1 time"> + </span>]:<span class="lineCov">          1 : export default (activeURL) =&gt; {</span></a>
<a name="5"><span class="lineNum">       5 </span>                :<span class="lineCov">          1 : </span></a>
<a name="6"><span class="lineNum">       6 </span>                :<span class="lineCov">          1 :         [</span></a>
<a name="7"><span class="lineNum">       7 </span>                :<span class="lineCov">          1 :                 'moz-blob',</span></a>
<a name="8"><span class="lineNum">       8 </span>                :<span class="lineCov">          1 :                 'moz-chunked-text',</span></a>
<a name="9"><span class="lineNum">       9 </span>                :<span class="lineCov">          1 :                 'moz-chunked-arraybuffer'</span></a>
<a name="10"><span class="lineNum">      10 </span>           [<span class="branchCov" title="Branch 0 was taken 3 times"> + </span>]:<span class="lineCov">          1 :         ].forEach((rt) =&gt; {</span></a>
<a name="11"><span class="lineNum">      11 </span>                :<span class="lineCov">          3 : </span></a>
<a name="12"><span class="lineNum">      12 </span>                :<span class="lineCov">          3 :                 let xhr = new XMLHttpRequest();</span></a>
<a name="13"><span class="lineNum">      13 </span>                :<span class="lineCov">          3 :                 xhr.responseType = rt;</span></a>
<a name="14"><span class="lineNum">      14 </span>                :<span class="lineCov">          3 : </span></a>
<a name="15"><span class="lineNum">      15 </span>                :<span class="lineCov">          3 :                 assert.strictEqual(xhr.responseType, '');</span></a>
<a name="16"><span class="lineNum">      16 </span>                :<span class="lineCov">          1 :         });</span></a>
<a name="17"><span class="lineNum">      17 </span>                :<span class="lineCov">          1 : }</span></a>
<a name="18"><span class="lineNum">      18 </span>                :<span class="lineCov">          1 : ﻿</span></a>
<a name="19"><span class="lineNum">      19 </span>                :<span class="lineCov">          1 : /*</span></a>
<a name="20"><span class="lineNum">      20 </span>                :<span class="lineCov">          1 :  * historical.html</span></a>
<a name="21"><span class="lineNum">      21 </span>                :<span class="lineCov">          1 :  *</span></a>
<a name="22"><span class="lineNum">      22 </span>                :<span class="lineCov">          1 : </span></a>
<a name="23"><span class="lineNum">      23 </span>                :<span class="lineCov">          1 : &lt;!doctype html&gt;</span></a>
<a name="24"><span class="lineNum">      24 </span>                :<span class="lineCov">          1 : &lt;meta charset=&quot;utf-8&quot;&gt;</span></a>
<a name="25"><span class="lineNum">      25 </span>                :<span class="lineCov">          1 : &lt;title&gt;Historical features&lt;/title&gt;</span></a>
<a name="26"><span class="lineNum">      26 </span>                :<span class="lineCov">          1 : &lt;script src=&quot;/resources/testharness.js&quot;&gt;&lt;/script&gt;</span></a>
<a name="27"><span class="lineNum">      27 </span>                :<span class="lineCov">          1 : &lt;script src=&quot;/resources/testharnessreport.js&quot;&gt;&lt;/script&gt;</span></a>
<a name="28"><span class="lineNum">      28 </span>                :<span class="lineCov">          1 : &lt;div id=&quot;log&quot;&gt;&lt;/div&gt;</span></a>
<a name="29"><span class="lineNum">      29 </span>                :<span class="lineCov">          1 : &lt;script&gt;</span></a>
<a name="30"><span class="lineNum">      30 </span>                :<span class="lineCov">          1 : [&quot;moz-blob&quot;, &quot;moz-chunked-text&quot;, &quot;moz-chunked-arraybuffer&quot;].forEach(function(rt) {</span></a>
<a name="31"><span class="lineNum">      31 </span>                :<span class="lineCov">          1 :   test(function() {</span></a>
<a name="32"><span class="lineNum">      32 </span>                :<span class="lineCov">          1 :     var xhr = new XMLHttpRequest();</span></a>
<a name="33"><span class="lineNum">      33 </span>                :<span class="lineCov">          1 :     xhr.responseType = rt;</span></a>
<a name="34"><span class="lineNum">      34 </span>                :<span class="lineCov">          1 :     assert_equals(xhr.responseType, &quot;&quot;);</span></a>
<a name="35"><span class="lineNum">      35 </span>                :<span class="lineCov">          1 :   }, &quot;Support for responseType = &quot; + rt);</span></a>
<a name="36"><span class="lineNum">      36 </span>                :<span class="lineCov">          1 : });</span></a>
<a name="37"><span class="lineNum">      37 </span>                :<span class="lineCov">          1 : &lt;/script&gt;</span></a>
<a name="38"><span class="lineNum">      38 </span>                :<span class="lineCov">          1 : </span></a>
<a name="39"><span class="lineNum">      39 </span>                :<span class="lineCov">          1 :  *</span></a>
<a name="40"><span class="lineNum">      40 </span>                :<span class="lineCov">          1 :  * historical.html</span></a>
<a name="41"><span class="lineNum">      41 </span>                :<span class="lineCov">          1 :  */</span></a>
</pre>
      </td>
    </tr>
  </table>
  <br>

  <table width="100%" border=0 cellspacing=0 cellpadding=0>
    <tr><td class="ruler"><img src="../../../glass.png" width=3 height=3 alt=""></td></tr>
    <tr><td class="versionInfo">Generated by: <a href="http://ltp.sourceforge.net/coverage/lcov.php" target="_parent">LCOV version 1.14</a></td></tr>
  </table>
  <br>

</body>
</html>
